Added serialization and communication packages

+++ c++/src/rtadp-proto/Worker2.cpp
@@ -20,7 +20,7 @@ Worker2::Worker2() : WorkerBase() {
             {"name": "assembly", "type": "string"},
             {"name": "name", "type": "string"},
             {"name": "serial_number", "type": "string"},
-            {"name": "timestamp", "type": "long"},
+            {"name": "timestamp", "type": "double"},
             {"name": "source_timestamp", "type": ["null", "long"]},
             {"name": "units", "type": "string"},
             {"name": "archive_suppress", "type": "boolean"},
